Some cleanings, removed VS2008 debúg compiler warnings: modified trunk/ctp2_code/gs/gameobj/bldque.cpp modified trunk/ctp2_code/gs/gameobj/CityData.cpp modified trunk/ctp2_code/gs/gameobj/Unit.cpp modified trunk/ctp2_code/gs/utility/TurnCnt.cpp modified trunk/ctp2_code/mapgen/Crater.cpp modified trunk/ctp2_code/ui/aui_ctp2/pattern.cpp modified trunk/ctp2_code/ui/aui_ctp2/radarmap.cpp modified trunk/ctp2_code/ui/interface/spnewgamediffscreen.cpp
git-svn-id: http://ctp2.darkdust.net/anonsvn/trunk@1037 34d2fc0c-33df-0310-8a34-85f8040ba873

+++ ctp2_code/ui/aui_ctp2/radarmap.cpp
@@ -1377,17 +1377,17 @@ void RadarMap::RedrawTile( const MapPoint *point )
                                    offsetpos.y + 1 + adjust
                                   );
 
-		dRect.left = sRect.left * (m_tilePixelWidth/2);
-		dRect.right = sRect.right * (m_tilePixelWidth/2);
-		dRect.top = sRect.top * m_tilePixelHeight;
-		dRect.bottom = sRect.bottom * m_tilePixelHeight;
+		dRect.left   = sRect.left   * static_cast<float>(m_tilePixelWidth/2);
+		dRect.right  = sRect.right  * static_cast<float>(m_tilePixelWidth/2);
+		dRect.top    = sRect.top    * static_cast<float>(m_tilePixelHeight);
+		dRect.bottom = sRect.bottom * static_cast<float>(m_tilePixelHeight);
 
 		primitives_Scale16(m_tempSurface, m_mapSurface, sRect, dRect, m_filter);
 
 		sRect.left = 0;
-		sRect.right = 1.0 + adjust;
+		sRect.right = 1.0f + adjust;
 		dRect.left = 0;
-		dRect.right = sRect.right * (m_tilePixelWidth/2);
+		dRect.right = sRect.right * static_cast<float>(m_tilePixelWidth/2);
 
 		primitives_Scale16(m_tempSurface, m_mapSurface, sRect, dRect, m_filter);
 	}
@@ -1403,10 +1403,10 @@ void RadarMap::RedrawTile( const MapPoint *point )
                                    offsetpos.y + 1 + adjust
                                   );
 		
-		dRect.left = sRect.left * (m_tilePixelWidth/2);
-		dRect.right = sRect.right * (m_tilePixelWidth/2);
-		dRect.top = sRect.top * m_tilePixelHeight;
-		dRect.bottom = sRect.bottom * m_tilePixelHeight;
+		dRect.left   = sRect.left   * static_cast<float>(m_tilePixelWidth/2);
+		dRect.right  = sRect.right  * static_cast<float>(m_tilePixelWidth/2);
+		dRect.top    = sRect.top    * static_cast<float>(m_tilePixelHeight);
+		dRect.bottom = sRect.bottom * static_cast<float>(m_tilePixelHeight);
 
 		primitives_Scale16(m_tempSurface, m_mapSurface, sRect, dRect, m_filter);
 	}
